DP solver passed first test
