+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/dormandPrince/Clas12Swim.java
@@ -1,0 +1,8 @@
+package cnuphys.dormandPrince;
+
+/*
+ * A set of static methods for CLAS12 swims
+ */
+public class Clas12Swim {
+
+}
