+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/dormandPrince/DormandPrince.java
@@ -1,4 +1,4 @@
-package cnuphys.dormondPrince;
+package cnuphys.dormandPrince;
 
 public class DormandPrince {
