+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/dormandPrince/ODE.java
@@ -1,4 +1,4 @@
-package cnuphys.dormondPrince;
+package cnuphys.dormandPrince;
 
 /**
  * Interface representing an ordinary differential equation (ODE).
