+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/dormandPrince/SwimmerODE.java
@@ -1,11 +1,14 @@
-package cnuphys.adaptiveSwim;
+package cnuphys.dormandPrince;
 
+import cnuphys.adaptiveSwim.AdaptiveSwimmer;
 import cnuphys.magfield.FieldProbe;
-import cnuphys.rk4.IDerivative;
-import cnuphys.swim.Swimmer;
 
-public class Derivative implements IDerivative {
-
+/**
+ * Thios is the ODE for swimming particles through a magnetic field
+ */
+ 
+public class SwimmerODE implements ODE {
+	
 	private FieldProbe _probe;
 
 	private double _momentum;  //GeV/c
@@ -18,6 +21,10 @@ public class Derivative implements IDerivative {
 	//for mag field result
 	float b[] = new float[3];
 
+	
+	// Speed of light in m/s
+	public static final double C = 299792458.0; // m/s
+	
 	/**
 	 * The derivative for swimming through a magnetic field
 	 * 
@@ -28,13 +35,14 @@ public class Derivative implements IDerivative {
 	 * @param field
 	 *            the magnetic field
 	 */
-	public Derivative(int charge, double momentum, FieldProbe field) {
+	public SwimmerODE(int charge, double momentum, FieldProbe field) {
 		_probe = field;
 		_momentum = momentum;
 //units of this  alpha are 1/(kG*m)
 		_alpha = 1.0e-10 * charge * AdaptiveSwimmer.C / _momentum;
 	}
-	
+
+
 
 	/**
 	 * Compute the derivatives given the value of s (path length) and the values
@@ -43,10 +51,11 @@ public class Derivative implements IDerivative {
 	 * @param s    the value of the independent variable path length (input).
 	 * @param u    the values of the state vector ([x,y,z, tx = px/p, ty = py/p, tz = pz/p]) at s
 	 *             (input).
-	 * @param du will be filled with the values of the derivatives at s (output).
+	 * @return the values of the derivatives at s (output).
 	 */
 	@Override
-	public void derivative(double s, double[] u, double[] du) {
+	public double[] getDerivatives(double s, double[] u) {
+		
 		double Bx = 0.0;
 		double By = 0.0;
 		double Bz = 0.0;
@@ -67,12 +76,14 @@ public class Derivative implements IDerivative {
 			Bz = b[2];
 		}
 
-		du[3] = _alpha * (u[4] * Bz - u[5] * By); // vyBz-vzBy
-		du[4] = _alpha * (u[5] * Bx - u[3] * Bz); // vzBx-vxBz
-		du[5] = _alpha * (u[3] * By - u[4] * Bx); // vxBy-vyBx
+		double du[] = new double[6];
 		du[0] = u[3];
 		du[1] = u[4];
 		du[2] = u[5];
+        du[3] = _alpha * (u[4] * Bz - u[5] * By); // vyBz-vzBy
+        du[4] = _alpha * (u[5] * Bx - u[3] * Bz); // vzBx-vxBz
+        du[5] = _alpha * (u[3] * By - u[4] * Bx); // vxBy-vyBx
+        return du;
 	}
 
 }
